fix: allow Threads view to show if MAT does not show it
In the event that MAT does not return detailed thread information, it
might return blank strings in the columns. Force-casting this to Boolean
results in an Exception. Instead of throwing an Exception, we attempt to
parse this to a Boolean which returns 'false' instead. It allows the
data to be presented as false, which is incorrect, but the API is
specified as `boolean`, so this is probably the best answer.